feat: add clerk authentication
