+++ src/app/(auth)/signup/[[...signup]]/page.tsx
@@ -1,0 +1,10 @@
+import { SignUp } from '@clerk/nextjs';
+import React from 'react';
+
+// type Props = {}
+
+const LoginPage: React.FC = (props) => {
+    return <SignUp />;
+};
+
+export default LoginPage;
