+++ src/app/(marketing)/page.tsx
@@ -1,0 +1,10 @@
+import HomeLoginButton from './components/HomeLoginButton';
+
+export default function Home() {
+    return (
+        <main className="flex min-h-screen flex-col items-center p-24">
+            Hello World!
+            <HomeLoginButton />
+        </main>
+    );
+}
